KEYFRAME ADDITION AND removal

+++ Lightning2/Animation/Types/AnimationPropertyFactory.cs
@@ -14,7 +14,7 @@
             int int1 = Convert.ToInt32(obj1);
             int int2 = Convert.ToInt32(obj2);
 
-            return int1 + ((int2 - int1) * Convert.ToInt32((double)cur / (double)max));
+            return int1 + ((int2 - int1) * Convert.ToInt32((double)cur / max));
         }
 
         internal static double GetDoubleValue(object obj1, object obj2, long cur, long max)
@@ -22,7 +22,7 @@
             double double1 = Convert.ToDouble(obj1);
             double double2 = Convert.ToDouble(obj2);
 
-            return double1 + ((double2 - double1) * ((double)cur / (double)max));
+            return double1 + ((double2 - double1) * ((double)cur / max));
         }
 
         internal static float GetFloatValue(object obj1, object obj2, long cur, long max)
@@ -30,7 +30,7 @@
             float float1 = Convert.ToSingle(obj1);
             float float2 = Convert.ToSingle(obj2);
 
-            return float1 + ((float2 - float1) * ((float)cur / (float)max));
+            return float1 + ((float2 - float1) * ((float)cur / max));
         }
 
         internal static Vector2 GetVector2Value(object obj1, object obj2, long cur, long max)
@@ -38,7 +38,7 @@
             Vector2 vec1 = (Vector2)obj1;
             Vector2 vec2 = (Vector2)obj2;
 
-            return vec1 + ((vec2 - vec1) * ((float)cur / (float)max));
+            return vec1 + ((vec2 - vec1) * ((float)cur / max));
         }
     }
 }
